requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect==1.2.0
distro==1.9.0
et-xmlfile==1.1.0
hiredis==2.3.2
idna==3.7
jmespath==1.0.1
lxml==5.2.1
more-itertools==10.2.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parsel==1.9.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==1.1.0
pytz==2024.1
redis==5.0.4
requests==2.31.0
six==1.16.0
tabula-py==2.9.0
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240425
types-redis==4.6.0.20240425
types-setuptools==70.0.0.20240524
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
xlrd==2.0.1
